refactor(orchestrator): nodes 를 노드별 모듈 6종으로 분할
1,714줄 → base 246 · orchestrator 145 · planner 418 · executor 518 ·
reviewer 77 · self_correction 376 · __init__ 28. 전부 800 이내.

정의 이동만 했고 본문은 한 글자도 바꾸지 않았다 (split_audit 으로 정의
7개 · 모듈레벨 할당 3개 전부 본문 일치, ruff format 이후 재실행도 0건).

## try/except ImportError 블록은 원자 단위다

RAG·MCP optional 의존 블록은 쪼갤 수 없다 — RAG_AVAILABLE 과
get_project_context 를 다른 모듈로 보내면 graceful degradation 구조가
깨진다. 쓰는 클래스를 따라 RAG 블록은 planner, MCP 블록은 executor 로
통째로 옮겼다(AST 역인덱스 실측: PlannerNode·ExecutorNode 만 사용).

RAG_AVAILABLE 은 실측상 **어느 클래스도 읽지 않는다**(미사용). 그래도
원본에 있으므로 보존하며, 같은 try 문장 안의 get_project_context 와
갈릴 수 없으므로 planner 로 간다.

## 순환 import 검증 — 이 대상의 세 번째 그물

except ImportError 는 **어떤** ImportError 든 삼키므로, 분할이 순환
import 를 만들면 플래그가 조용히 False 가 되고 fallback 이 빈 문자열을
돌려준다. 이 상태는 ruff·mypy 와 무관하고 split_audit 도 0건이다
(블록 텍스트가 동일하다). api/usage 에서는 route_table 이 그 자리였지만
nodes 에는 라우트가 없다.

분할 전후 실측이 완전히 일치한다:
  RAG_AVAILABLE=True · MCP_AVAILABLE=True ·
  get_project_context.__module__='services.rag_service' ·
  MCPToolExecutor=orchestrator.tools.MCPToolExecutor

## 테스트 문자열 패치 7회 갱신

계획서 예상과 정확히 일치(7회, 잔여 0건). 타깃은 '그 이름을 읽는
클래스가 사는 모듈'이다:
  record_usage_best_effort · LLMService  → nodes.base     (BaseNode)
  audit_task_status_change · AuditService → nodes.executor (ExecutorNode)

앞의 둘은 모듈 지역이라 재노출로 살아나지 않고, 뒤의 둘은 클래스 속성이라
재노출로도 동작한다. 그럼에도 넷 다 실제 사용처 경로로 맞췄다 — __init__
재노출을 소비자 요구분(노드 클래스 6종)으로 좁게 유지해야 갱신 누락이
AttributeError 로 즉시 드러난다(B5 Task 2 교훈).

Red-Green: base 경로를 분할 전 형태로 되돌리면 2 failed, 복원하면 32 passed.

게이트: ruff 0 · ruff format 0 · mypy 310파일 0 ·
pytest 1 failed, 1362 passed, 2 skipped
(Task 2 후 1360 + optional deps 2건, 실패 1건은 알려진 RAG 플레이크)
